# 24001200-km7-ald-BasicBankingSystem-ch5/src/docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Basic Banking Simulation API",
    "description": "This is an API for the challenged issued by Binar Academy, using Swagger as a form of API Documentation for ease of use",
    "contact": {
      "name": "Aldy Ramadhan Syahputra"
    },
    "license": {
      "name": "Free"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090/api/v1"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get All Users",
        "description": "Getting All Users with no exceptions",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully Retrieved All Data"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "name": {
                            "type": "string",
                            "example": "Aldy Ramadhan Syahputra"
                          },
                          "email": {
                            "type": "string",
                            "example": "dyrmdhns101@gmail.com"
                          },
                          "profile": {
                            "type": "object",
                            "properties": {
                              "identity_type": {
                                "type": "string",
                                "example": "KTP"
                              },
                              "identity_number": {
                                "type": "string",
                                "example": "64646464"
                              },
                              "address": {
                                "type": "string",
                                "example": "Jln Perjuangan"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create a User",
        "description": "Creates a new user with the provided details",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "email",
                  "name",
                  "password",
                  "profile"
                ],
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Carl Vincenzo"
                  },
                  "email": {
                    "type": "string",
                    "example": "carl123@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "12345678"
                  },
                  "profile": {
                    "required": [
                      "address",
                      "identity_number",
                      "identity_type"
                    ],
                    "type": "object",
                    "properties": {
                      "identity_type": {
                        "type": "string",
                        "example": "KTP"
                      },
                      "identity_number": {
                        "type": "string",
                        "example": "887726"
                      },
                      "address": {
                        "type": "string",
                        "example": "Jln. Pangeran"
                      }
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User Created Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "User Created Successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 8
                        },
                        "name": {
                          "type": "string",
                          "example": "Carl Vincenzo"
                        },
                        "email": {
                          "type": "string",
                          "example": "carl123@gmail.com"
                        },
                        "password": {
                          "type": "string",
                          "example": "12345678"
                        },
                        "profile": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "example": 8
                            },
                            "user_id": {
                              "type": "integer",
                              "example": 8
                            },
                            "identity_type": {
                              "type": "string",
                              "example": "KTP"
                            },
                            "identity_number": {
                              "type": "string",
                              "example": "887726"
                            },
                            "address": {
                              "type": "string",
                              "example": "Jln. Pangeran"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed response cases",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "message": {
                      "type": "string",
                      "enum": [
                        "Failed To Create User, Make Sure To fill all Form",
                        "Email must be a Gmail address",
                        "Password must be at least 8 characters long",
                        "Identity number must be a number",
                        "Email Already Used",
                        "Identity Number Already Used"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User by ID",
        "description": "Get a single user by their ID",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully Retrieved The Data"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "Aldy Ramadhan Syahputra"
                        },
                        "email": {
                          "type": "string",
                          "example": "dyrmdhns101@gmail.com"
                        },
                        "profile": {
                          "type": "object",
                          "properties": {
                            "identity_type": {
                              "type": "string",
                              "example": "KTP"
                            },
                            "identity_number": {
                              "type": "string",
                              "example": "64646464"
                            },
                            "address": {
                              "type": "string",
                              "example": "Jln Perjuangan"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "message": {
                      "type": "string",
                      "example": "There's no Such User!"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}